+++ backend/src/main/java/com/abdi/cardiscover/CarDiscover.java
@@ -2,8 +2,7 @@ package com.abdi.cardiscover;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -13,6 +12,8 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
 @ComponentScan(basePackages = "com.abdi.cardiscover")
 @EnableJpaRepositories
 @EnableAutoConfiguration
+@CrossOrigin(origins = {"http://192.168.1.196:4200", "http://localhost:8080"})
+
 public class CarDiscover extends SpringBootServletInitializer {
