+++ backend/src/main/java/com/abdi/cardiscover/controller/Supplier.java
@@ -1,0 +1,5 @@
+package com.abdi.cardiscover.controller;
+
+public class Supplier {
+    
+}
