--- backend/src/main/java/com/abdi/cardiscover/controller/reservation/GetReservation.java
@@ -1,5 +1,0 @@
-package com.abdi.cardiscover.controller.reservation;
-
-public class GetReservation {
-    
-}
