--- backend/src/main/java/com/abdi/cardiscover/controller/speciality/CreateSpeciality.java
@@ -1,5 +1,0 @@
-package com.abdi.cardiscover.controller.speciality;
-
-public class CreateSpeciality {
-    
-}
