--- backend/src/main/java/com/abdi/cardiscover/controller/supplier/CreateSupplier.java
@@ -1,5 +1,0 @@
-package com.abdi.cardiscover.controller.supplier;
-
-public class CreateSupplier {
-    
-}
